+++ DICOM/Network/DicomPresentationContextCollection.cs
@@ -19,7 +19,12 @@ namespace Dicom.Network {
 			if (_pc.Count == 0)
 				return 1;
 
-			return (byte)(_pc.Max(x => x.Key) + 2);
+			var id = _pc.Max(x => x.Key) + 2;
+
+			if (id >= 256)
+				throw new DicomNetworkException("Too many presentation contexts configured for this association!");
+
+			return (byte)id;
 		}
 
 		public void Add(DicomUID abstractSyntax, params DicomTransferSyntax[] transferSyntaxes) {
@@ -38,7 +43,14 @@ namespace Dicom.Network {
 		public void AddFromRequest(DicomRequest request) {
 			if (request is DicomCStoreRequest) {
 				var cstore = request as DicomCStoreRequest;
-				var pc = _pc.Values.FirstOrDefault(x => x.AbstractSyntax == request.SOPClassUID && x.AcceptedTransferSyntax == cstore.TransferSyntax);
+
+				var pcs = _pc.Values.Where(x => x.AbstractSyntax == request.SOPClassUID);
+				if (cstore.TransferSyntax == DicomTransferSyntax.ImplicitVRLittleEndian)
+					pcs = _pc.Values.Where(x => x.GetTransferSyntaxes().Contains(DicomTransferSyntax.ImplicitVRLittleEndian));
+				else
+					pcs = _pc.Values.Where(x => x.AcceptedTransferSyntax == cstore.TransferSyntax);
+
+				var pc = pcs.FirstOrDefault();
 				if (pc == null) {
 					var tx = new List<DicomTransferSyntax>();
 					if (cstore.TransferSyntax != DicomTransferSyntax.ImplicitVRLittleEndian)